project1_rag_production/requirements.txt
# ==============================================================================
# RAG Production Pipeline - Requirements
# ==============================================================================
# Production-ready RAG with evaluation frameworks
# Python 3.9-3.11 compatible

# ==============================================================================
# Core Deep Learning (Install separately with CUDA)
# ==============================================================================
# pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu124

transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
safetensors>=0.4.0

# ==============================================================================
# Embeddings
# ==============================================================================
sentence-transformers>=2.2.2

# ==============================================================================
# Vector Search
# ==============================================================================
faiss-cpu>=1.7.4
chromadb>=0.4.0

# ==============================================================================
# Rerankers (Multiple Options)
# ==============================================================================
# Cohere Reranker API (best quality, paid)
cohere>=4.0.0

# Jina Reranker (best open-source) - uses sentence-transformers

# BGE Reranker - uses sentence-transformers

# OpenAI for LLM Reranker
openai>=1.0.0

# Anthropic for Claude Reranker (optional)
anthropic>=0.18.0

# ==============================================================================
# LangChain (for ContextualCompressionRetriever)
# ==============================================================================
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-cohere>=0.1.0

# ==============================================================================
# Evaluation Frameworks
# ==============================================================================
ragas>=0.1.0
deepeval>=0.20.0

# ==============================================================================
# MLflow & Tracking
# ==============================================================================
mlflow>=2.9.0
tensorboard>=2.15.0

# ==============================================================================
# Visualization
# ==============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# ==============================================================================
# Data Processing
# ==============================================================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
pyarrow>=14.0.0

# ==============================================================================
# Text Processing
# ==============================================================================
nltk>=3.8
tiktoken>=0.5.0
rank-bm25>=0.2.2

# ==============================================================================
# Configuration
# ==============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0
omegaconf>=2.3.0

# ==============================================================================
# Utilities
# ==============================================================================
tqdm>=4.65.0
requests>=2.31.0
tenacity>=8.2.0
colorama>=0.4.6

# ==============================================================================
# Model Optimization
# ==============================================================================
bitsandbytes>=0.41.0
optimum>=1.14.0

# ==============================================================================
# Testing
# ==============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0

# ==============================================================================
# Code Quality
# ==============================================================================
black>=23.0.0
ruff>=0.1.0

# ==============================================================================
# Additional
# ==============================================================================
httpx>=0.25.0
pydantic>=2.0.0
typing-extensions>=4.8.0
